added EventCard component
